Update part.cfg
This should address the missing LF/OX from the BBQ probe.  As it was, in the editor and r-click popups, there was no fuel for the fire extinguisher showing, and it wouldn't run when attached to the bottom directly and launched.  the resources were 'outside' the "part" module due to bracketing error.